requirements.txt
jinja2
flask
pillow
numpy
pandas
open3d
torch
trimesh
transformers
tensorflow
keras
scikit-learn
xgboost